requirement.txt
# -----------------------------
# FastAPI Core
# -----------------------------
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
python-dotenv==1.0.1

# -----------------------------
# Security / Rate Limiting
# -----------------------------
slowapi==0.1.9

# -----------------------------
# LLM + Quantization
# -----------------------------
transformers==4.38.2
torch==2.2.1
accelerate==0.27.2
bitsandbytes==0.42.0

# -----------------------------
# Embeddings
# -----------------------------
sentence-transformers==2.6.1

# -----------------------------
# Vector Database
# -----------------------------
chromadb==0.4.24

# -----------------------------
# PDF Processing
# -----------------------------
pypdf==4.0.2

# -----------------------------
# Token Utilities
# -----------------------------
tiktoken==0.6.0